Move the patchouli stuff
