+++ src/main/java/forestry/core/patchouli/processor/package-info.java
@@ -1,0 +1,10 @@
+@ParametersAreNonnullByDefault
+@FieldsAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+package forestry.core.patchouli.processor;
+
+import javax.annotation.ParametersAreNonnullByDefault;
+
+import forestry.core.utils.FieldsAreNonnullByDefault;
+
+import mcp.MethodsReturnNonnullByDefault;
